adding imageButtonPage into its own component
